Revert "Move TimestampedMessage comparison functions to Records util"
This reverts commit 0e3630b8db61c1b04376cc0b0cad8db88f666be6.
