+++ src/core/message.ts
@@ -1,5 +1,5 @@
 import type { SignatureInput } from '../jose/jws/general/types.js';
-import type { BaseDecodedAuthorizationPayload, BaseMessage, Descriptor } from './types.js';
+import type { BaseDecodedAuthorizationPayload, BaseMessage, Descriptor, TimestampedMessage } from './types.js';
 
 import { computeCid } from '../utils/cid.js';
 import type { GeneralJws } from '../jose/jws/general/types.js';
@@ -140,5 +140,54 @@ export abstract class Message<M extends BaseMessage> {
     const signer = await GeneralJwsSigner.create(authPayloadBytes, [signatureInput]);
 
     return signer.getJws();
+  }
+
+
+  /**
+   * @returns newest message in the array. `undefined` if given array is empty.
+   */
+  public static async getNewestMessage(messages: TimestampedMessage[]): Promise<TimestampedMessage | undefined> {
+    let currentNewestMessage: TimestampedMessage | undefined = undefined;
+    for (const message of messages) {
+      if (currentNewestMessage === undefined || await Message.isNewer(message, currentNewestMessage)) {
+        currentNewestMessage = message;
+      }
+    }
+
+    return currentNewestMessage;
+  }
+
+  /**
+   * Checks if first message is newer than second message.
+   * @returns `true` if `a` is newer than `b`; `false` otherwise
+   */
+  public static async isNewer(a: TimestampedMessage, b: TimestampedMessage): Promise<boolean> {
+    const aIsNewer = (await Message.compareModifiedTime(a, b) > 0);
+    return aIsNewer;
+  }
+
+  /**
+   * Checks if first message is older than second message.
+   * @returns `true` if `a` is older than `b`; `false` otherwise
+   */
+  public static async isOlder(a: TimestampedMessage, b: TimestampedMessage): Promise<boolean> {
+    const aIsNewer = (await Message.compareModifiedTime(a, b) < 0);
+    return aIsNewer;
+  }
+
+  /**
+   * Compares the `dateModified` of the given messages with a fallback to message CID according to the spec.
+   * @returns 1 if `a` is larger/newer than `b`; -1 if `a` is smaller/older than `b`; 0 otherwise (same age)
+   */
+  public static async compareModifiedTime(a: TimestampedMessage, b: TimestampedMessage): Promise<number> {
+    if (a.descriptor.dateModified > b.descriptor.dateModified) {
+      return 1;
+    } else if (a.descriptor.dateModified < b.descriptor.dateModified) {
+      return -1;
+    }
+
+    // else `dateModified` is the same between a and b
+    // compare the `dataCid` instead, the < and > operators compare strings in lexicographical order
+    return Message.compareCid(a, b);
   }
 }
